ISSUE-16 - Remove cargo core usage: * Remove cargo core usage from scan module * Remove cargo core usage from traversal module * Remove cargo core usage from format module * Remove cargo core usage from scan module * Refactor krates_utils into separate modules * Update tarpaulin version https://github.com/xd009642/tarpaulin/issues/517
